code/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
decorator==5.1.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
py==1.11.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
requests==2.32.3
retry==0.9.2
rich==13.8.1
six==1.16.0
urllib3==2.2.3
